In theory receiving all other visible users within a radius around a position from the webservice should be working. Implementation of the webservice has been completed but not been tested yet.

+++ exercise_2/code/app/src/main/java/de/hhn/aib/swlab/wise1920/group01/exercise2/model/sync/UserAPI.java
@@ -31,6 +31,6 @@ public interface UserAPI {
     @GET("location/{radius}/{lat}/{long}")
     Call<List<MapObjectDummy>> getEverythingAround(@Header("Authorization") String jwt,
                                                    @Path("radius") Integer radius,
-                                                   @Path("lat") Integer latitude,
-                                                   @Path("long") Integer longitude);
+                                                   @Path("lat") Double latitude,
+                                                   @Path("long") Double longitude);
 }
